[Cleanup] Rename _CPlugin_SensorTypeHelper.h to _Plugin_SensorTypeHelper.h
It is meant for plugins, not controllers, so using the CPlugin prefix was confusing.
