+++ src/_Plugin_Helper.h
@@ -25,7 +25,7 @@
 #include "src/Helpers/ESPEasy_time_calc.h"
 #include "src/Helpers/I2C_access.h"
 #include "src/Helpers/StringGenerator_GPIO.h"
-#include "src/Helpers/_CPlugin_SensorTypeHelper.h"
+#include "src/Helpers/_Plugin_SensorTypeHelper.h"
 #include "src/Helpers/_Plugin_Helper_serial.h"
 
 #include "src/WebServer/HTML_wrappers.h"
