+++ src/src/Helpers/_Plugin_SensorTypeHelper.h
