+++ src/src/WebServer/ControllerPage.cpp
@@ -18,7 +18,7 @@
 #include "../Globals/Settings.h"
 
 #include "../Helpers/_CPlugin_Helper_webform.h"
-#include "../Helpers/_CPlugin_SensorTypeHelper.h"
+#include "../Helpers/_Plugin_SensorTypeHelper.h"
 #include "../Helpers/ESPEasy_Storage.h"
 #include "../Helpers/StringConverter.h"
